Created user model && jwt authentication set up

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 // @desc get goals
 // @route GET /api/goals
 // @access PRIVATE
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     
     res.status(200).json(goals)
 
@@ -23,7 +24,8 @@ const setGoals = asyncHandler(async (req, res) => {
     };
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -38,6 +40,22 @@ const updateGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found!')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found!')
+    }
+    
+    //Make sure the logged in user matches the goal user
+if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authoriezd')
+}
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true
     } )
@@ -54,6 +72,22 @@ const deleteGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found!')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found!')
+    }
+    
+    //Make sure the logged in user matches the goal user
+if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authoriezd')
+}
+
     const deledGoal = await Goal.findByIdAndDelete(req.params.id)
     res.status(200).json({id: req.params.id})
 })
